study/src/test/java/local/algorithm/leetcode/BinaryTreeHasPathSumIIITest.java
package local.algorithm.leetcode;

import local.algorithm.leetcode.common.TreeNode;
import org.testng.annotations.Test;

import static org.testng.Assert.assertEquals;

/**
 * @author cuixiaoshuang
 * @date 2020-01-06
 **/
public class BinaryTreeHasPathSumIIITest {

    @Test
    public void testPathSum() {
        TreeNode root = new TreeNode(10,
                new TreeNode(5,
                        new TreeNode(3,
                                new TreeNode(3),
                                new TreeNode(-2)),
                        new TreeNode(2,
                                null,
                                new TreeNode(1))),
                new TreeNode(-3,
                        null,
                        new TreeNode(11)));

        int result = new BinaryTreeHasPathSumIII().pathSum(root, 8);
        assertEquals(result, 3);
    }
}